requirements.txt
Flask==3.0.0
gunicorn==21.2.0
face-recognition==1.3.0
Pillow==10.1.0
numpy==1.24.3
